+++ src/main/kotlin/ar/edu/unq/desapp/grupoL/criptop2p/webservice/PublisherRestService.kt
@@ -27,7 +27,7 @@ class PublisherRestService {
         var response : ResponseEntity<*>?
         try {
 
-            val publicacion = publisherService!!.publicar(id.toLong(),entity)
+            val publicacion = publisherService!!.publicar(id,entity)
 
             ResponseEntity.status(200)
             response = ResponseEntity.ok().body(publicacion)
